move user context one directory up and use the existing CurrentUser type
